Broke Blocker, Main Button Check Functional
Was unable to test display-related state data from my dashboard tests because state was being held in the top level component and shared with only display. By holding state in dashboard and passing it to display from there, I was able to test my dashboard functions' affect on display-related data.